requirements.txt
Django
djangorestframework
django-filter
drf-nested-routers
